requirements.txt
Flask
gunicorn
playwright
twocaptcha-python